Replaced use of old settings components
