+++ packages/serverless-logic-web-tools/src/editor/EditorToolbar.tsx
@@ -71,7 +71,6 @@ import {
 import { useCancelableEffect } from "@kie-tools-core/react-hooks/dist/useCancelableEffect";
 import { useGitHubAuthInfo } from "../settings/github/Hooks";
 import { AuthStatus, GithubScopes, useSettings, useSettingsDispatch } from "../settings/SettingsContext";
-import { SettingsTabs } from "../settings/SettingsModalBody";
 import { FileLabel } from "../workspace/components/FileLabel";
 import { WorkspaceLabel } from "../workspace/components/WorkspaceLabel";
 import { PromiseStateWrapper } from "@kie-tools-core/react-hooks/dist/PromiseState";
@@ -94,6 +93,8 @@ import { WorkspaceStatusIndicator } from "../workspace/components/WorkspaceStatu
 import { WorkspaceKind } from "@kie-tools-core/workspaces-git-fs/dist/worker/api/WorkspaceOrigin";
 import { useEditorEnvelopeLocator } from "../envelopeLocator/EditorEnvelopeLocatorContext";
 import { UrlType, useImportableUrl } from "../workspace/hooks/ImportableUrlHooks";
+import { Link } from "react-router-dom";
+import { routes } from "../navigation/Routes";
 
 export interface Props {
   alerts: AlertsController | undefined;
@@ -727,10 +728,12 @@ If you are, it means that creating this Gist failed and it can safely be deleted
               {!canPushToGitRepository && (
                 <>
                   <Divider />
-                  <DropdownItem onClick={() => settingsDispatch.open(SettingsTabs.GITHUB)}>
-                    <Button isInline={true} variant={ButtonVariant.link}>
-                      Configure GitHub token...
-                    </Button>
+                  <DropdownItem>
+                    <Link to={routes.settings.github.path({})}>
+                      <Button isInline={true} variant={ButtonVariant.link}>
+                        Configure GitHub token...
+                      </Button>
+                    </Link>
                   </DropdownItem>
                 </>
               )}
@@ -1072,7 +1075,7 @@ If you are, it means that creating this Gist failed and it can safely be deleted
                 )}
 
                 {!canPushToGitRepository && (
-                  <AlertActionLink onClick={() => settingsDispatch.open(SettingsTabs.GITHUB)}>
+                  <AlertActionLink onClick={() => history.push(routes.settings.github.path({}))}>
                     {`Configure GitHub token...`}
                   </AlertActionLink>
                 )}
@@ -1603,7 +1606,7 @@ If you are, it means that creating this Gist failed and it can safely be deleted
                                   {!canPushToGitRepository && (
                                     <>
                                       <Divider />
-                                      <DropdownItem onClick={() => settingsDispatch.open(SettingsTabs.GITHUB)}>
+                                      <DropdownItem onClick={() => history.push(routes.settings.github.path({}))}>
                                         <Button isInline={true} variant={ButtonVariant.link}>
                                           Configure GitHub token...
                                         </Button>
@@ -1661,7 +1664,7 @@ If you are, it means that creating this Gist failed and it can safely be deleted
                                   {!canPushToGitRepository && (
                                     <>
                                       <Divider />
-                                      <DropdownItem onClick={() => settingsDispatch.open(SettingsTabs.GITHUB)}>
+                                      <DropdownItem onClick={() => history.push(routes.settings.github.path({}))}>
                                         <Button isInline={true} variant={ButtonVariant.link}>
                                           Configure GitHub token...
                                         </Button>
@@ -1752,7 +1755,7 @@ export function PushToGitHubAlertActionLinks(props: {
   kind?: WorkspaceKind;
   remoteRef?: string;
 }) {
-  const settingsDispatch = useSettingsDispatch();
+  const history = useHistory();
 
   if (props.kind === WorkspaceKind.GIT && !props.remoteRef) {
     throw new Error("Should specify remoteRef for GIT workspaces");
@@ -1761,7 +1764,7 @@ export function PushToGitHubAlertActionLinks(props: {
   return (
     <>
       {!props.canPush && (
-        <AlertActionLink onClick={() => settingsDispatch.open(SettingsTabs.GITHUB)}>
+        <AlertActionLink onClick={() => history.push(routes.settings.github.path({}))}>
           {`Configure GitHub token...`}
         </AlertActionLink>
       )}
