+++ packages/serverless-logic-web-tools/src/kieSandboxExtendedServices/KieSandboxExtendedServicesIcon.tsx
@@ -24,17 +24,19 @@ import { ConnectedIcon } from "@patternfly/react-icons/dist/js/icons/connected-i
 import { DisconnectedIcon } from "@patternfly/react-icons/dist/js/icons/disconnected-icon";
 import { useKieSandboxExtendedServices } from "./KieSandboxExtendedServicesContext";
 import { useSettingsDispatch } from "../settings/SettingsContext";
-import { SettingsTabs } from "../settings/SettingsModalBody";
 import { useAppI18n } from "../i18n";
+import { routes } from "../navigation/Routes";
+import { useHistory } from "react-router";
 
 export function KieSandboxExtendedServicesIcon() {
   const kieSandboxExtendedServices = useKieSandboxExtendedServices();
   const { i18n } = useAppI18n();
   const settingsDispatch = useSettingsDispatch();
+  const history = useHistory();
 
   const toggleKieSandboxExtendedServices = useCallback(() => {
     if (kieSandboxExtendedServices.status === KieSandboxExtendedServicesStatus.RUNNING) {
-      settingsDispatch.open(SettingsTabs.KIE_SANDBOX_EXTENDED_SERVICES);
+      history.push(routes.settings.kie_sandbox_extended_services.path({}));
     }
 
     if (!kieSandboxExtendedServices.outdated) {
