+++ packages/serverless-logic-web-tools/src/settings_bak/SettingsContext.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,25 +14,30 @@
  * limitations under the License.
  */
 
-import { OpenShiftConnection } from "@kie-tools-core/openshift/dist/service/OpenShiftConnection";
-import { OpenShiftService } from "@kie-tools-core/openshift/dist/service/OpenShiftService";
-import { Octokit } from "@octokit/rest";
 import * as React from "react";
-import { useContext, useEffect, useMemo, useState } from "react";
+import { useCallback, useContext, useEffect, useMemo, useState } from "react";
 import { getCookie, setCookie } from "../cookies";
-import { SwfServiceCatalogStore } from "../editor/api/SwfServiceCatalogStore";
-import { useKieSandboxExtendedServices } from "../kieSandboxExtendedServices/KieSandboxExtendedServicesContext";
-import { KieSandboxExtendedServicesStatus } from "../kieSandboxExtendedServices/KieSandboxExtendedServicesStatus";
-import { OpenShiftInstanceStatus } from "../openshift/OpenShiftInstanceStatus";
-import { FeaturePreviewSettingsConfig, readFeaturePreviewConfigCookie } from "./featurePreview/FeaturePreviewConfig";
-import { GITHUB_AUTH_TOKEN_COOKIE_NAME } from "./github/GitHubSettings";
-import { KafkaSettingsConfig, readKafkaConfigCookie } from "./kafka/KafkaSettingsConfig";
+import { Octokit } from "@octokit/rest";
+import { useQueryParams } from "../queryParams/QueryParamsContext";
+import { Modal, ModalVariant } from "@patternfly/react-core/dist/js/components/Modal";
+import { SettingsModalBody, SettingsTabs } from "./SettingsModalBody";
 import { readOpenShiftConfigCookie } from "./openshift/OpenShiftSettingsConfig";
+import { OpenShiftConnection } from "@kie-tools-core/openshift/dist/service/OpenShiftConnection";
+import { OpenShiftInstanceStatus } from "../openshift/OpenShiftInstanceStatus";
+import { OpenShiftService } from "@kie-tools-core/openshift/dist/service/OpenShiftService";
+import { useHistory, useRouteMatch } from "react-router";
+import { QueryParams, routes } from "../navigation/Routes";
+import { GITHUB_AUTH_TOKEN_COOKIE_NAME } from "./github/GitHubSettingsTab";
+import { KafkaSettingsConfig, readKafkaConfigCookie } from "./kafka/KafkaSettingsConfig";
 import { readServiceAccountConfigCookie, ServiceAccountSettingsConfig } from "./serviceAccount/ServiceAccountConfig";
 import {
   readServiceRegistryConfigCookie,
   ServiceRegistrySettingsConfig,
 } from "./serviceRegistry/ServiceRegistryConfig";
+import { useKieSandboxExtendedServices } from "../kieSandboxExtendedServices/KieSandboxExtendedServicesContext";
+import { KieSandboxExtendedServicesStatus } from "../kieSandboxExtendedServices/KieSandboxExtendedServicesStatus";
+import { SwfServiceCatalogStore } from "../editor/api/SwfServiceCatalogStore";
+import { FeaturePreviewSettingsConfig, readFeaturePreviewConfigCookie } from "./featurePreview/FeaturePreviewConfig";
 
 export enum AuthStatus {
   SIGNED_OUT,
@@ -64,6 +69,8 @@ export class ExtendedServicesConfig {
 }
 
 export interface SettingsContextType {
+  isOpen: boolean;
+  activeTab: SettingsTabs;
   openshift: {
     status: OpenShiftInstanceStatus;
     config: OpenShiftConnection;
@@ -92,6 +99,8 @@ export interface SettingsContextType {
 }
 
 export interface SettingsDispatchContextType {
+  open: (activeTab?: SettingsTabs) => void;
+  close: () => void;
   openshift: {
     service: OpenShiftService;
     setStatus: React.Dispatch<React.SetStateAction<OpenShiftInstanceStatus>>;
@@ -123,6 +132,32 @@ export const SettingsContext = React.createContext<SettingsContextType>({} as an
 export const SettingsDispatchContext = React.createContext<SettingsDispatchContextType>({} as any);
 
 export function SettingsContextProvider(props: any) {
+  const queryParams = useQueryParams();
+  const history = useHistory();
+  const [isOpen, setOpen] = useState(false);
+  const [activeTab, setActiveTab] = useState(SettingsTabs.GITHUB);
+  const isRouteInSettingsSection = useRouteMatch(routes.settings.home.path({}));
+
+  useEffect(() => {
+    setOpen(queryParams.has(QueryParams.SETTINGS));
+    setActiveTab((queryParams.get(QueryParams.SETTINGS) as SettingsTabs) ?? SettingsTabs.GITHUB);
+  }, [queryParams]);
+
+  const open = useCallback(
+    (activeTab = SettingsTabs.GITHUB) => {
+      history.replace({
+        search: queryParams.with(QueryParams.SETTINGS, activeTab).toString(),
+      });
+    },
+    [history, queryParams]
+  );
+
+  const close = useCallback(() => {
+    history.replace({
+      search: queryParams.without(QueryParams.SETTINGS).toString(),
+    });
+  }, [history, queryParams]);
+
   //github
   const [githubAuthStatus, setGitHubAuthStatus] = useState(AuthStatus.LOADING);
   const [githubOctokit, setGitHubOctokit] = useState<Octokit>(new Octokit());
@@ -222,6 +257,8 @@ export function SettingsContextProvider(props: any) {
 
   const dispatch = useMemo(() => {
     return {
+      open,
+      close,
       openshift: {
         service: openshiftService,
         setStatus: setOpenshiftStatus,
@@ -249,8 +286,10 @@ export function SettingsContextProvider(props: any) {
       },
     };
   }, [
+    close,
     githubAuthService,
     githubOctokit,
+    open,
     openshiftService,
     kieSandboxExtendedServices.saveNewConfig,
     serviceCatalogStore,
@@ -258,6 +297,8 @@ export function SettingsContextProvider(props: any) {
 
   const value = useMemo(() => {
     return {
+      isOpen,
+      activeTab,
       openshift: {
         status: openshiftStatus,
         config: openshiftConfig,
@@ -285,6 +326,8 @@ export function SettingsContextProvider(props: any) {
       },
     };
   }, [
+    isOpen,
+    activeTab,
     openshiftStatus,
     openshiftConfig,
     githubAuthStatus,
@@ -300,7 +343,19 @@ export function SettingsContextProvider(props: any) {
 
   return (
     <SettingsContext.Provider value={value}>
-      <SettingsDispatchContext.Provider value={dispatch}>{props.children}</SettingsDispatchContext.Provider>
+      <SettingsDispatchContext.Provider value={dispatch}>
+        {(isRouteInSettingsSection || githubAuthStatus !== AuthStatus.LOADING) && <>{props.children}</>}
+        <Modal
+          title="Settings"
+          isOpen={!isRouteInSettingsSection && isOpen}
+          onClose={close}
+          variant={ModalVariant.large}
+        >
+          <div style={{ height: "calc(100vh * 0.5)" }} className={"kie-tools--settings-modal-content"}>
+            <SettingsModalBody />
+          </div>
+        </Modal>
+      </SettingsDispatchContext.Provider>
     </SettingsContext.Provider>
   );
 }
