+++ packages/serverless-logic-web-tools/src/settings_bak/openshift/OpenShiftSettingsTabSimpleConfig.tsx
